Implemented Base Working Model
I linked the front end and back end to create what is at this point the base model of this application. You can select between Chicago and London and when selected, the respective and current weather data is shown. The scope is limited and UI is ugly but it works!