Added and using ControlsFX
I added ControlsFX to the projects buildpath. This handy library has one feature I want to use in my project which is the autocomplete feature. I want to be able to search up cities using some sort of autocomplete and intended to build my own with a Trie. However, doing this would have been difficult if not for this library. No need to reinvent the wheel as they say.